backend/requirements.txt
# ============================================================
#  SmartTask — Backend Requirements
#  Tested for: Python 3.14 on Windows (64-bit)
#  Install with: pip install -r requirements.txt
# ============================================================

Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0

# mysqlclient 2.2.8 has a pre-built .whl for Python 3.14 Windows (cp314-cp314-win_amd64)
mysqlclient==2.2.8

# Pillow 12.1.1 has a pre-built .whl for Python 3.14 Windows (cp314-cp314-win_amd64)
Pillow==12.1.1